refactor: extract rate limiter module
